+++ ForestManagement/changelog.md
@@ -1,5 +1,10 @@
 ﻿# Changelog
 
+## 1.5.73 (2024-01-12)
+
+- Upd: Schema - added parameters "MayBeContainedIn" and "MustBeContainedIn" to support both options
+- Fix: Schema - empty admin descriptions cause a change, even though they are configured as empty.
+
 ## 1.5.71 (2023-11-02)
 
 - Fix: Schema Ldif - fails to deploy changes with custom credentials